move pt 1
